refactored defaultvault classes
