+++ larky/src/main/java/com/verygood/security/larky/modules/vgs/vault/defaults/AliasGenerator.java
@@ -1,0 +1,7 @@
+package com.verygood.security.larky.modules.vgs.vault.defaults;
+
+import net.starlark.java.eval.EvalException;
+
+interface AliasGenerator {
+    String tokenize(String value) throws EvalException;
+}
