+++ larky/src/main/java/com/verygood/security/larky/modules/vgs/vault/defaults/UUIDAliasGenerator.java
@@ -1,0 +1,10 @@
+package com.verygood.security.larky.modules.vgs.vault.defaults;
+
+import net.starlark.java.eval.EvalException;
+
+class UUIDAliasGenerator extends RawAliasGenerator {
+    @Override
+    public String tokenize(String value) throws EvalException {
+        return String.format("tok_%s", super.tokenize(value)).substring(0, 30);
+    }
+}
