+++ larky/src/test/resources/META-INF/services/com.verygood.security.larky.modules.vgs.vault.spi.LarkyVault
@@ -1,1 +1,1 @@
-com.verygood.security.larky.modules.vgs.vault.DefaultVault
+com.verygood.security.larky.modules.vgs.vault.defaults.DefaultVault
